refactor(files): Remove unnecessary files, group all card related components
