+++ app/components/Card/CardFront.tsx
