+++ app/components/Card/MachineTypeLabel.tsx
